moved models to subfolder, updated imports, tested with pytest and played around on localhost
