+++ api/holiday_acres_api/models/Users.py
@@ -1,0 +1,11 @@
+from django.db import models
+from django.db.models.deletion import SET_DEFAULT, SET_NULL
+
+
+class User(models.Model):
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+    # USER DATA
+    first_name = models.CharField(max_length=200)
+    last_name = models.CharField(max_length=200)
+    email = models.CharField(max_length=200)
