+++ api/holiday_acres_api/tests/horse_unit_test.py
@@ -1,8 +1,8 @@
 import pytest
 
-from holiday_acres_api.models import Horse
-from holiday_acres_api.models import Paddock
-from holiday_acres_api.models import User
+from holiday_acres_api.models.Users import User
+from holiday_acres_api.models.Paddocks import Paddock
+from holiday_acres_api.models.Horses import Horse
 
 
 """
