+++ api/holiday_acres_api/tests/user_unit_test.py
@@ -1,7 +1,6 @@
 import pytest
 
-from holiday_acres_api.models import User
-
+from holiday_acres_api.models.Users import User
 
 """
 Can create and fetch basic user in DB
