+++ api/holiday_acres_api/views.py
@@ -1,11 +1,13 @@
 from django.shortcuts import render
 from rest_framework import viewsets
-from holiday_acres_api.serializers import UserSerializer
-from holiday_acres_api.serializers import PaddockSerializer
-from holiday_acres_api.serializers import HorseSerializer
-from holiday_acres_api.models import User
-from holiday_acres_api.models import Paddock
-from holiday_acres_api.models import Horse
+from holiday_acres_api.serializers import (
+    UserSerializer,
+    PaddockSerializer,
+    HorseSerializer,
+)
+from holiday_acres_api.models.Users import User
+from holiday_acres_api.models.Paddocks import Paddock
+from holiday_acres_api.models.Horses import Horse
 
 
 class UserViewSet(viewsets.ModelViewSet):
